[ chore ] 패키지 버전별 분류 구조 철회에 따른 패키지 이동 ( #21 )
