+++ src/main/java/com/qring/review/application/message/KafkaMessageProducerV1.java
@@ -1,4 +1,4 @@
-package com.qring.review.application.v1.message;
+package com.qring.review.application.message;
 
 public interface KafkaMessageProducerV1 {
