+++ src/main/java/com/qring/review/application/service/ReservationServiceV1.java
@@ -1,7 +1,7 @@
-package com.qring.review.application.v1.service;
+package com.qring.review.application.service;
 
 import com.qring.review.application.global.dto.ResDTO;
-import com.qring.review.application.v1.res.ReservationGetByIdResDTOV1;
+import com.qring.review.application.res.ReservationGetByIdResDTOV1;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
