+++ src/main/java/com/qring/review/application/service/RestaurantServiceV1.java
@@ -1,4 +1,4 @@
-package com.qring.review.application.v1.service;
+package com.qring.review.application.service;
 
 import com.qring.review.application.global.dto.ResDTO;
 import org.springframework.http.ResponseEntity;
