+++ src/main/java/com/qring/review/application/service/ReviewServiceV1.java
@@ -1,20 +1,20 @@
-package com.qring.review.application.v1.service;
+package com.qring.review.application.service;
 
 import com.qring.review.application.global.exception.DuplicateResourceException;
 import com.qring.review.application.global.exception.EntityNotFoundException;
 import com.qring.review.application.global.exception.UnauthorizedAccessException;
-import com.qring.review.application.v1.message.KafkaMessageProducerV1;
-import com.qring.review.application.v1.message.ReviewEventMessageDTOV1;
-import com.qring.review.application.v1.res.ReservationGetByIdResDTOV1;
-import com.qring.review.application.v1.res.ReviewGetByIdResDTOV1;
-import com.qring.review.application.v1.res.ReviewPostResDTOV1;
-import com.qring.review.application.v1.res.ReviewSearchResDTOV1;
+import com.qring.review.application.message.KafkaMessageProducerV1;
+import com.qring.review.application.message.ReviewEventMessageDTOV1;
+import com.qring.review.application.res.ReservationGetByIdResDTOV1;
+import com.qring.review.application.res.ReviewGetByIdResDTOV1;
+import com.qring.review.application.res.ReviewPostResDTOV1;
+import com.qring.review.application.res.ReviewSearchResDTOV1;
 import com.qring.review.domain.model.ReviewEntity;
 import com.qring.review.domain.model.constraint.ReservationStatus;
 import com.qring.review.domain.repository.ReviewRepository;
 import com.qring.review.infrastructure.util.PassportUtil;
-import com.qring.review.presentation.v1.req.PostReviewReqDTOV1;
-import com.qring.review.presentation.v1.req.PutReviewReqDTOV1;
+import com.qring.review.presentation.req.PostReviewReqDTOV1;
+import com.qring.review.presentation.req.PutReviewReqDTOV1;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
