+++ src/main/java/com/qring/review/infrastructure/messaging/KafkaMessageProducerImplV1.java
@@ -1,7 +1,7 @@
 package com.qring.review.infrastructure.messaging;
 
-import com.qring.review.application.v1.message.KafkaMessageProducerV1;
-import com.qring.review.application.v1.message.ReviewEventMessageDTOV1;
+import com.qring.review.application.message.KafkaMessageProducerV1;
+import com.qring.review.application.message.ReviewEventMessageDTOV1;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
