+++ src/main/java/com/qring/review/infrastructure/repository/JpaReviewRepository.java
@@ -1,10 +1,7 @@
 package com.qring.review.infrastructure.repository;
 
-import com.qring.review.application.v1.message.ReviewStatisticsDTOV1;
 import com.qring.review.domain.model.ReviewEntity;
-import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
